% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{check_col}
\alias{check_col}
\title{Check col}
\usage{
check_col(DT = NULL, col = NULL, arg = NULL, extra = NULL)
}
\arguments{
\item{DT}{data.table}

\item{col}{column name}

\item{arg}{argument name}

\item{extra}{extras}
}
\description{
Check col
}
\keyword{internal}
