% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{degdeviation}
\alias{degdeviation}
\title{Degree Deviation}
\usage{
degdeviation(DT, degree, id, splitBy = NULL)
}
\arguments{
\item{DT}{}

\item{degree}{}

\item{id}{}

\item{splitBy}{this is not the splitBy (eg season)}
}
\value{
Column added named degdev
}
\description{
Degree Deviation
}
