% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{layer_correlation}
\alias{layer_correlation}
\title{Layer correlation}
\usage{
layer_correlation(gLs, attr = "weight")
}
\arguments{
\item{gLs}{Expects a list of two weighted graphs}

\item{attr}{Graph attribute to use, default assuming the weight is stored.}
}
\description{
Layer correlation
}
