% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{layer_eigen}
\alias{layer_eigen}
\title{Calculate eigenvector centrality for each graph in a list}
\usage{
layer_eigen(graphLs)
}
\arguments{
\item{graphLs}{}
}
\description{
Calculate eigenvector centrality for each graph in a list
}
