% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{layer_neighbors}
\alias{layer_neighbors}
\title{Neighbourhood}
\usage{
layer_neighbors(DT, id, var = NULL, splitBy = NULL)
}
\arguments{
\item{DT}{a data.table with column "group" generated by spatsoc::group_pts}

\item{id}{}

\item{var}{if there are subgroups within the DT, eg if the DT is the result of some kind of sensitivity analysis on spatsoc spatial buffers, include that field here}

\item{splitBy}{this is not the splitBy (eg season)}
}
\value{
The input DT with additional column "neighborhood" and optionally "splitNeighorhood" if a column was provided for the `splitBy` argument.
}
\description{
Number of neighbors adjacent to each actor. Calculated excluding self from set of neighbors.
}
\examples{
# Load data.table and spatsoc
library(data.table)
library(spatsoc)

# Read example data
DT <- fread(system.file("extdata", "DT.csv", package = "spatsoc"))

# Cast the character column to POSIXct
DT[, datetime := as.POSIXct(datetime, tz = 'UTC')]

# Temporal grouping
group_times(DT, datetime = 'datetime', threshold = '20 minutes')

# Spatial grouping with timegroup
group_pts(DT, threshold = 5, id = 'ID',
          coords = c('X', 'Y'), timegroup = 'timegroup')

layer_neighbors(DT, 'id', var, splitBy = 'season')
}
