% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{layer_relevance}
\alias{layer_relevance}
\title{Relevance}
\usage{
layer_relevance(DT, id, splitBy)
}
\arguments{
\item{DT}{}

\item{id}{}

\item{splitBy}{}

\item{var}{}
}
\value{

}
\description{
Proportion of neighbours present on each layer.
}
\references{
Berlingerio, Michele, et al. "Foundations of multidimensional network analysis." 2011 international conference on advances in social networks analysis and mining. IEEE, 2011.
}
