% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{layer_similarity}
\alias{layer_similarity}
\title{Layer similarity}
\usage{
layer_similarity(matrices, pattern, splitBy)
}
\arguments{
\item{matrices}{property matrices generated using `property_matrix`}

\item{pattern}{}

\item{splitBy}{column indicating which groups to compare ensuring each group of rows = 2.}
}
\value{

}
\description{
Layer similarity
}
