% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{layer_similarity_ordinal}
\alias{layer_similarity_ordinal}
\title{Layer similarity - ordinal}
\usage{
layer_similarity_ordinal(matrices, pattern, splitBy)
}
\value{

}
\description{
Layer similarity - ordinal
}
