% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{list_gbi}
\alias{list_gbi}
\title{GBI}
\usage{
list_gbi(DT, id, splitList, splitBy, group = "group")
}
\arguments{
\item{DT}{}

\item{id}{}

\item{splitList}{}

\item{splitBy}{}

\item{group}{}
}
\description{
GBI
}
