% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{list_graphs}
\alias{list_graphs}
\title{Graphs}
\usage{
list_graphs(netLs, mode = "undirected", diag = FALSE,
  weighted = TRUE)
}
\arguments{
\item{netLs}{}

\item{mode}{}

\item{diag}{}

\item{weighted}{}
}
\description{
Graphs
}
