% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{neigh}
\alias{neigh}
\title{Neighbourhood}
\usage{
neigh(DT, id, splitBy = NULL)
}
\description{
Number of neighbors adjacent to each actor. Calculated excluding self from set of neighbors.
}
\examples{
neigh(DT, 'id', splitBy = 'season')
}
