% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{redundancy}
\alias{redundancy}
\title{Connective Redundancy}
\usage{
redundancy(DT)
}
\arguments{
\item{DT}{}
}
\description{
When connective redundancy is 0, all edges on all layers 
are necessary to preserve the social ties
}
\details{
DT must be result of `neigh()` function, having columns "splitNeighborhood" 
and "neighborhood"
}
