% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{relevance}
\alias{relevance}
\title{Relevance}
\usage{
relevance(DT, id, var, splitBy)
}
\arguments{
\item{DT}{}

\item{id}{}

\item{var}{}

\item{splitBy}{}
}
\description{
Proportion of neighbours present on each layer.
}
\references{
Berlingerio, Michele, et al. "Foundations of multidimensional network analysis." 2011 international conference on advances in social networks analysis and mining. IEEE, 2011.
}
