% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{shear_xy}
\alias{shear_xy}
\title{Shear XY for stacked plotting
Thanks to [@rafapereirabr](https://github.com/rafapereirabr) for this gist (https://gist.github.com/rafapereirabr/97a7c92d40f91cd20a10e8e0165a0aef) and Barry Rowlingson for the original SO answer (http://gis.stackexchange.com/questions/189490/plot-tilted-map-in-r)}
\usage{
shear_xy(DT, coordcols, shearmatrix = matrix(c(2, 1.2, 0, 1), 2, 2))
}
\arguments{
\item{DT}{}

\item{coordcols}{length 2}

\item{shearmatrix}{}
}
\description{
Shear XY for stacked plotting
Thanks to [@rafapereirabr](https://github.com/rafapereirabr) for this gist (https://gist.github.com/rafapereirabr/97a7c92d40f91cd20a10e8e0165a0aef) and Barry Rowlingson for the original SO answer (http://gis.stackexchange.com/questions/189490/plot-tilted-map-in-r)
}
