% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{multi_degree}
\alias{multi_degree}
\title{Multidegree}
\usage{
multi_degree(DT, degree, id, splitBy = NULL)
}
\arguments{
\item{DT}{}

\item{degree}{}

\item{id}{}

\item{splitBy}{this is not the splitBy (eg season)}
}
\value{
Column added named multideg
}
\description{
Multidegree
}
