% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{property_matrix}
\alias{property_matrix}
\title{Property Matrix}
\usage{
property_matrix(DT, id, layer, metric)
}
\arguments{
\item{DT}{}

\item{id}{}

\item{layer}{}

\item{metric}{}
}
\description{
Property Matrix
}
\references{
Bródka P, Chmiel A,Magnani M, Ragozini G. 2018 Quantifying layer
similarity in multiplex networks: a systematic study. R.Soc.opensci. 5:171747.
http://dx.doi.org/10.1098/rsos.171747
}
